relation with users

+++ controllers/bankMasterController.js
@@ -3,8 +3,9 @@ const BankMaster = require('../models/bankMasterSchema');
 // Create
 exports.createBankDetails = async (req, res) => {
   try {
+    const userId = req.userId;
     const { bankName, ifscCode, branch } = req.body;
-    const newBankDetails = await BankMaster.create({ bankName, ifscCode, branch });
+    const newBankDetails = await BankMaster.create({ bankName, ifscCode, branch, userId });
     res.status(201).json(newBankDetails);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -14,7 +15,8 @@ exports.createBankDetails = async (req, res) => {
 // Read all
 exports.getBankDetails = async (req, res) => {
   try {
-    const bankDetails = await BankMaster.findAll();
+    const userId = req.userId;
+    const bankDetails = await BankMaster.findAll({ where: { userId } });
     res.json(bankDetails);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -24,14 +26,15 @@ exports.getBankDetails = async (req, res) => {
 // Update
 exports.updateBankDetails = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
     const { bankName, ifscCode, branch } = req.body;
-    const bankDetails = await BankMaster.findByPk(id);
+    const bankDetails = await BankMaster.findOne({ where: { id, userId } });
     if (!bankDetails) return res.status(404).json({ error: "Builder not found" });
 
-    bankDetails.bankName=bankName
-    bankDetails.ifscCode=ifscCode
-    bankDetails.branch=branch
+    bankDetails.bankName = bankName
+    bankDetails.ifscCode = ifscCode
+    bankDetails.branch = branch
     await bankDetails.save();
 
     res.json(bankDetails);
@@ -43,8 +46,9 @@ exports.updateBankDetails = async (req, res) => {
 // Delete
 exports.deleteBankDetails = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
-    const deleted = await BankMaster.destroy({ where: { id } });
+    const deleted = await BankMaster.destroy({ where: { id, userId } });
     if (!deleted) return res.status(404).json({ error: "Bank Details not found" });
     res.json({ message: "Bank Details deleted successfully" });
   } catch (err) {
