+++ controllers/builderMasterController.js
@@ -3,8 +3,9 @@ const BuilderMaster = require('../models/builderMasterSchema');
 // Create
 exports.createBuilder = async (req, res) => {
   try {
+    const userId = req.userId;
     const { builderMaster } = req.body;
-    const newBuilder = await BuilderMaster.create({ builderMaster});
+    const newBuilder = await BuilderMaster.create({ builderMaster,userId});
     res.status(201).json(newBuilder);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -14,8 +15,9 @@ exports.createBuilder = async (req, res) => {
 // Read all
 exports.getBuilders = async (req, res) => {
   try {
-    const builder = await BuilderMaster.findAll();
-    res.json(builder);
+    const userId = req.userId;
+    const builders = await BuilderMaster.findAll({ where: { userId } });
+    res.json(builders);
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -24,9 +26,10 @@ exports.getBuilders = async (req, res) => {
 // Update
 exports.updateBuilder = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
     const { builderMaster } = req.body;
-    const builder = await BuilderMaster.findByPk(id);
+    const builder = await BuilderMaster.findOne({ where: {id, userId } });
     if (!builder) return res.status(404).json({ error: "Builder not found" });
 
    
@@ -42,8 +45,9 @@ exports.updateBuilder = async (req, res) => {
 // Delete
 exports.deleteBuilder = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
-    const deleted = await BuilderMaster.destroy({ where: { id } });
+    const deleted = await BuilderMaster.destroy({ where: { id,userId } });
     if (!deleted) return res.status(404).json({ error: "Builder not found" });
     res.json({ message: "Builder deleted successfully" });
   } catch (err) {
