+++ controllers/departmentMasterController.js
@@ -4,8 +4,9 @@ const DepartmentMaster = require("../models/departmentMasterSchema");
 
 exports.createDepartmentDetails = async (req, res) => {
     try {
+        const userId = req.userId;
         const { departmentName } = req.body;
-        const newDepartment = await DepartmentMaster.create({ departmentName });
+        const newDepartment = await DepartmentMaster.create({ departmentName,userId });
         res.status(201).json(newDepartment);
     } catch (error) {
         res.status(500).json({ error: error.message })
@@ -16,7 +17,8 @@ exports.createDepartmentDetails = async (req, res) => {
 
 exports.getDepartmentDetails = async (req, res) => {
     try {
-        const departments = await DepartmentMaster.findAll();
+        const userId = req.userId;
+        const departments = await DepartmentMaster.findAll({ where: { userId } });
         res.json(departments);
     } catch (error) {
         res.status(500).json({ error: error.message });
@@ -27,9 +29,10 @@ exports.getDepartmentDetails = async (req, res) => {
 
 exports.updateDepartmentDetails = async (req, res) => {
     try {
+        const userId = req.userId;
         const { departmentID } = req.params;
         const { departmentName } = req.body;
-        const department = await DepartmentMaster.findByPk(departmentID);
+        const department = await DepartmentMaster.findOne({ where: {departmentID, userId } });
         if (!department) return res.status(404).json({ error: "Department not found" });
 
         department.departmentName = departmentName;
@@ -45,8 +48,9 @@ exports.updateDepartmentDetails = async (req, res) => {
 
 exports.deleteDepartmentDetails = async (req, res) => {
     try {
+        const userId = req.userId;
         const { departmentID } = req.params;
-        const deleted = await DepartmentMaster.destroy({ where: { departmentID } });
+        const deleted = await DepartmentMaster.destroy({ where: { departmentID,userId } });
         if (!deleted) return res.status(404).json({ error: "Department not found" });
         res.json({ message: "Department deleted successfully" });
     } catch (error) {
