+++ controllers/expenseCategoryMasterController.js
@@ -3,8 +3,9 @@ const ExpenseCategoryMaster = require('../models/expenseCategoryMasterSchema');
 // Create
 exports.createExpenseCategory= async (req, res) => {
   try {
+    const userId = req.userId;
     const { expenseCategory, expenseHead } = req.body;
-    const newExpenseCategory= await ExpenseCategoryMaster.create({ expenseCategory, expenseHead });
+    const newExpenseCategory= await ExpenseCategoryMaster.create({ expenseCategory, expenseHead,userId });
     res.status(201).json(newExpenseCategory);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -14,7 +15,8 @@ exports.createExpenseCategory= async (req, res) => {
 // Read all
 exports.getExpenseCategorys = async (req, res) => {
   try {
-    const expenseCategorys = await ExpenseCategoryMaster.findAll();
+    const userId = req.userId;
+    const expenseCategorys = await ExpenseCategoryMaster.findAll({ where: { userId } });
     res.json(expenseCategorys);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -24,9 +26,10 @@ exports.getExpenseCategorys = async (req, res) => {
 // Update
 exports.updateExpenseCategory= async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
     const { expenseCategory, expenseHead } = req.body;
-    const expenseCategorys = await ExpenseCategoryMaster.findByPk(id);
+    const expenseCategorys = await ExpenseCategoryMaster.findOne({ where: {id, userId } });
     if (!expenseCategorys) return res.status(404).json({ error: "ExpenseCategorynot found" });
 
     expenseCategorys.expenseCategory = expenseCategory;
@@ -42,8 +45,9 @@ exports.updateExpenseCategory= async (req, res) => {
 // Delete
 exports.deleteExpenseCategory= async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
-    const deleted = await ExpenseCategoryMaster.destroy({ where: { id } });
+    const deleted = await ExpenseCategoryMaster.destroy({ where: { id,userId } });
     if (!deleted) return res.status(404).json({ error: "ExpenseCategorynot found" });
     res.json({ message: "ExpenseCategory deleted successfully" });
   } catch (err) {
