+++ controllers/fundPurposeController.js
@@ -3,8 +3,9 @@ const FundPurpose = require('../models/fundPurposeSchema');
 // Create
 exports.createFundPurpose = async (req, res) => {
   try {
+    const userId = req.userId;
     const { fundPurpose } = req.body;
-    const newFundPurpose = await FundPurpose.create({ fundPurpose});
+    const newFundPurpose = await FundPurpose.create({ fundPurpose,userId});
     res.status(201).json(newFundPurpose);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -14,7 +15,8 @@ exports.createFundPurpose = async (req, res) => {
 // Read all
 exports.getFundPurposes = async (req, res) => {
   try {
-    const fundPurpose = await FundPurpose.findAll();
+    const userId = req.userId;
+    const fundPurpose = await FundPurpose.findAll({ where: { userId } });
     res.json(fundPurpose);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -24,9 +26,10 @@ exports.getFundPurposes = async (req, res) => {
 // Update
 exports.updateFundPurpose = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
     const { fundPurpose } = req.body;
-    const fundPurposes = await FundPurpose.findByPk(id);
+    const fundPurposes = await FundPurpose.findOne({ where: {id, userId } });
     if (!fundPurposes) return res.status(404).json({ error: "FundPurpose not found" });
 
    
@@ -42,8 +45,9 @@ exports.updateFundPurpose = async (req, res) => {
 // Delete
 exports.deleteFundPurpose = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
-    const deleted = await FundPurpose.destroy({ where: { id } });
+    const deleted = await FundPurpose.destroy({ where: { id,userId } });
     if (!deleted) return res.status(404).json({ error: "Fund Purpose not found" });
     res.json({ message: "Fund Purpose deleted successfully" });
   } catch (err) {
