+++ controllers/fundSourceController.js
@@ -3,8 +3,9 @@ const FundSource = require('../models/fundSourceSchema');
 // Create
 exports.createFundSource = async (req, res) => {
   try {
+    const userId = req.userId;
     const { fundSource } = req.body;
-    const newFundSource = await FundSource.create({ fundSource});
+    const newFundSource = await FundSource.create({ fundSource,userId});
     res.status(201).json(newFundSource);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -14,7 +15,8 @@ exports.createFundSource = async (req, res) => {
 // Read all
 exports.getFundSources = async (req, res) => {
   try {
-    const fundSource = await FundSource.findAll();
+    const userId = req.userId;
+    const fundSource = await FundSource.findAll({ where: { userId } });
     res.json(fundSource);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -24,9 +26,10 @@ exports.getFundSources = async (req, res) => {
 // Update
 exports.updateFundSource = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
     const { fundSource } = req.body;
-    const fundSources = await FundSource.findByPk(id);
+    const fundSources = await FundSource.findOne({ where: {id, userId } });
     if (!fundSources) return res.status(404).json({ error: "FundSource not found" });
 
    
@@ -42,8 +45,9 @@ exports.updateFundSource = async (req, res) => {
 // Delete
 exports.deleteFundSource = async (req, res) => {
   try {
+    const userId = req.userId;
     const { id } = req.params;
-    const deleted = await FundSource.destroy({ where: { id } });
+    const deleted = await FundSource.destroy({ where: { id ,userId} });
     if (!deleted) return res.status(404).json({ error: "Fund Source not found" });
     res.json({ message: "Fund Source deleted successfully" });
   } catch (err) {
